fix: svg-props.tsx 컴포넌트화 (nav, tab 수정)
