+++ src/shared/assets/index.ts
@@ -1,0 +1,1 @@
+export * from './svg-props';
