+++ src/shared/assets/svg-props/FullLogoAlyacNoTextIcon.tsx
@@ -1,0 +1,7 @@
+export const FullLogoAlyacNoTextIcon = (
+  props: React.SVGProps<SVGSVGElement>,
+): React.JSX.Element => (
+  <svg viewBox="0 0 97 153" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+    <image x="0" y="0" width="97" height="153" href="/full-logo-alyac-no-text.webp" />
+  </svg>
+);
