+++ src/shared/assets/svg-props/FullLogoAlyacPngIcon.tsx
@@ -1,0 +1,5 @@
+export const FullLogoAlyacPngIcon = (props: React.SVGProps<SVGSVGElement>): React.JSX.Element => (
+  <svg viewBox="0 0 199 220" fill="none" xmlns="http://www.w3.org/2000/svg" {...props}>
+    <image x="0" y="0" width="199" height="220" href="/full-logo-alyac-png.webp" />
+  </svg>
+);
