+++ src/shared/assets/svg-props/MoonIcon.tsx
@@ -1,0 +1,15 @@
+export const MoonIcon = (props: React.SVGProps<SVGSVGElement>) => (
+  <svg
+    width="40"
+    height="20"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    strokeWidth="2"
+    strokeLinecap="round"
+    strokeLinejoin="round"
+    {...props}
+  >
+    <path d="M20.985 12.486a9 9 0 1 1-9.473-9.472c.405-.022.617.46.402.803a6 6 0 0 0 8.268 8.268c.344-.215.825-.004.803.401" />
+  </svg>
+);
