+++ src/shared/assets/svg-props/UploadImageIcon.tsx
@@ -1,0 +1,11 @@
+export const UploadImageIcon = (props: React.SVGProps<SVGSVGElement>): React.JSX.Element => (
+  <svg
+    viewBox="0 0 110 110"
+    fill="none"
+    xmlns="http://www.w3.org/2000/svg"
+    xmlnsXlink="http://www.w3.org/1999/xlink"
+    {...props}
+  >
+    <image x="0" y="0" width="110" height="110" href="/upload-image.webp" />
+  </svg>
+);
