+++ src/shared/assets/svg-props/types.ts
@@ -1,0 +1,6 @@
+export interface IconProps {
+  size?: number;
+  color?: string;
+  className?: string;
+  active?: boolean;
+}
