+++ src/widgets/top-basic-nav/ui/TopBasicNav.tsx
@@ -3,6 +3,7 @@ import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import { removeToken } from '@/entities/auth/lib/token';
+import { BackIcon, MonitorIcon, MoonIcon, MoreIcon, SunIcon } from '@/shared/assets';
 import { useTheme } from '@/shared/lib/theme';
 import { LogoutModal } from '@/shared/ui/modal';
 
@@ -28,61 +29,9 @@ export function TopBasicNav({ onSettings }: TopBasicNavProps) {
   };
 
   const themeLabel = {
-    light: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="40"
-        height="20"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <circle cx="12" cy="12" r="4" />
-        <path d="M12 2v2" />
-        <path d="M12 20v2" />
-        <path d="m4.93 4.93 1.41 1.41" />
-        <path d="m17.66 17.66 1.41 1.41" />
-        <path d="M2 12h2" />
-        <path d="M20 12h2" />
-        <path d="m6.34 17.66-1.41 1.41" />
-        <path d="m19.07 4.93-1.41 1.41" />
-      </svg>
-    ),
-    dark: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="40"
-        height="20"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <path d="M20.985 12.486a9 9 0 1 1-9.473-9.472c.405-.022.617.46.402.803a6 6 0 0 0 8.268 8.268c.344-.215.825-.004.803.401" />
-      </svg>
-    ),
-    system: (
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        width="40"
-        height="20"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <rect width="20" height="14" x="2" y="3" rx="2" />
-        <line x1="8" x2="16" y1="21" y2="21" />
-        <line x1="12" x2="12" y1="17" y2="21" />
-      </svg>
-    ),
+    light: <SunIcon />,
+    dark: <MoonIcon />,
+    system: <MonitorIcon />,
   };
 
   return (
@@ -92,15 +41,7 @@ export function TopBasicNav({ onSettings }: TopBasicNavProps) {
           onClick={() => navigate(-1)}
           className="text-foreground flex h-[32px] w-[32px] items-center justify-center rounded-md transition-colors hover:bg-gray-100"
         >
-          <svg width="24" height="24" viewBox="0 0 24 24" fill="none">
-            <path
-              d="M19 12H5M5 12L12 19M5 12L12 5"
-              stroke="currentColor"
-              strokeWidth="1.8"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            />
-          </svg>
+          <BackIcon />
         </button>
 
         <div className="relative">
@@ -108,11 +49,7 @@ export function TopBasicNav({ onSettings }: TopBasicNavProps) {
             onClick={() => setIsMenuOpen((prev) => !prev)}
             className="text-foreground flex h-[32px] w-[32px] items-center justify-center rounded-md transition-colors hover:bg-gray-100"
           >
-            <svg width="24" height="24" viewBox="0 0 24 24" fill="none">
-              <circle cx="12" cy="5" r="1.2" fill="currentColor" />
-              <circle cx="12" cy="12" r="1.2" fill="currentColor" />
-              <circle cx="12" cy="19" r="1.2" fill="currentColor" />
-            </svg>
+            <MoreIcon />
           </button>
 
           {isMenuOpen && (
