+++ src/widgets/top-upload-nav/ui/top-upload-nav.tsx
@@ -1,5 +1,7 @@
 import { useNavigate } from 'react-router-dom';
 
+import { BackIcon } from '@/shared/assets';
+
 interface TopUploadNavProps {
   label?: string;
   disabled?: boolean;
@@ -15,15 +17,7 @@ export function TopUploadNav({ label = '업로드', disabled = false, onSubmit }
         onClick={() => navigate(-1)}
         className="text-foreground flex h-[32px] w-[32px] items-center justify-center rounded-md transition-colors hover:bg-gray-100"
       >
-        <svg width="24" height="24" viewBox="0 0 24 24" fill="none">
-          <path
-            d="M19 12H5M5 12L12 19M5 12L12 5"
-            stroke="currentColor"
-            strokeWidth="1.8"
-            strokeLinecap="round"
-            strokeLinejoin="round"
-          />
-        </svg>
+        <BackIcon />
       </button>
 
       <button
